pyproject.toml
[project]
name = "regwatch"
version = "0.1.0"
description = "Regulatory Watcher for Union Investment Luxembourg S.A."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "jinja2>=3.1",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "sqlite-vec>=0.1.6",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "typer>=0.12",
  "httpx>=0.27",
  "feedparser>=6.0",
  "SPARQLWrapper>=2.0",
  "trafilatura>=1.9",
  "lxml_html_clean>=0.1",
  "pdfplumber>=0.11",
  "pypdf>=4.2",
  "python-dateutil>=2.9",
  "langdetect>=1.0.9",
  "langchain-text-splitters>=0.2",
  "apscheduler>=3.10",
  "python-slugify>=8.0",
  "tiktoken>=0.7",
  "python-multipart>=0.0.9",
  "beautifulsoup4>=4.12",
  "playwright>=1.40",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-httpx>=0.30",
  "ruff>=0.4",
  "mypy>=1.10",
  "freezegun>=1.5",
  "reportlab>=4.1",
]

[project.scripts]
regwatch = "regwatch.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["regwatch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "live: live network tests, excluded from default runs",
]
addopts = "-m 'not live'"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]

[tool.mypy]
python_version = "3.11"
strict = true
